Update the rendering fix to include various other SpecialPoint tiles By request of builder Jun
